requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
email-validator==1.1.2
SQLAlchemy==1.3.20
PyJWT==2.0.1
